model/ remote/mode/devices y ui/speaker/ Cambie todos los speaker para que usen float en el volumen, y lo hice andar en SpeakerScreen

+++ app/src/main/java/com/hci/TP3_HCI/model/Speaker.kt
@@ -8,7 +8,7 @@ class Speaker(
     id: String?,
     name: String,
     val status: Status,
-    val volume: Int,
+    val volume: Float,
     val genre: String,
     val song: SpeakerSong,
 ) : Device(id, name, DeviceType.SPEAKER) {
